+++ src/solver/variable/include/antares/solver/variable/economy/renewableGeneration.h
@@ -73,7 +73,7 @@ struct VCardRenewableGeneration
     //! Decimal precision
     static constexpr uint8_t decimal =   0;
     //! Number of columns used by the variable (One ResultsType per column)
-    static constexpr uint8_t columnCount =   9;
+    static constexpr int columnCount =    9;
     //! The Spatial aggregation
     static constexpr uint8_t spatialAggregate =   Category::spatialAggregateSum;
     static constexpr uint8_t spatialAggregateMode =   Category::spatialAggregateEachYear;
