+++ src/solver/variable/include/antares/solver/variable/setofareas.h
@@ -61,7 +61,7 @@ struct VCardAllSetsOfAreas
     //! Indentation (GUI)
     static constexpr uint8_t nodeDepthForGUI =   +1;
     //! Number of columns used by the variable (One ResultsType per column)
-    static constexpr uint8_t columnCount =   0;
+    static constexpr int columnCount =    0;
     //! The Spatial aggregation
     static constexpr uint8_t spatialAggregate =   Category::noSpatialAggregate;
     //! Intermediate values
